[domain] move HashModified into module
