+++ domain/src/main/scala/net/kemitix/s3thorp/domain/HashModified.scala
@@ -1,6 +1,4 @@
-package net.kemitix.s3thorp
-
-import net.kemitix.s3thorp.domain.{LastModified, MD5Hash}
+package net.kemitix.s3thorp.domain
 
 final case class HashModified(hash: MD5Hash,
                               modified: LastModified) {
